+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.3.0 (unreleased)
 - Fix gofakeitPrice return value
 - Use environment variables directly in hcl [#63](https://github.com/restbeast/restbeast/issues/63)
+- Added `unixTimestamp` function  [#64](https://github.com/restbeast/restbeast/issues/64)
 
 1.2.0
  - Fix crash on empty slice